app/requirements.txt
--extra-index-url https://pypi.org/simple

# Web framework
fastapi
uvicorn[standard]
pillow

# Vision + embeddings
torch
torchvision
sentence-transformers

# Optional but recommended for robustness
numpy
